Adding unit tests and PushPrice endpoint

+++ ReservationManagementSystem.Application/Features/Rooms/Commands/CreateRoom/CreateRoomRequest.cs
@@ -4,4 +4,4 @@ using ReservationManagementSystem.Application.Wrappers;
 
 namespace ReservationManagementSystem.Application.Features.Rooms.Commands.CreateRoom;
 
-public sealed record CreateRoomRequest(Guid RoomTypeId, string Number, int Floor, string? Note) : IRequest<Result<RoomResponse>>;
+public sealed record CreateRoomRequest(Guid RoomTypeId, string Number, byte Floor, string? Note) : IRequest<Result<RoomResponse>>;
