+++ ReservationManagementSystem.Application/Features/Rooms/Commands/UpdateRoom/UpdateRoomRequest.cs
@@ -4,4 +4,4 @@ using ReservationManagementSystem.Application.Wrappers;
 
 namespace ReservationManagementSystem.Application.Features.Rooms.Commands.UpdateRoom;
 
-public sealed record UpdateRoomRequest(Guid Id, Guid RoomTypeId, string Number, int Floor, string? Note) : IRequest<Result<RoomResponse>>;
+public sealed record UpdateRoomRequest(Guid Id, Guid RoomTypeId, string Number, byte Floor, string? Note) : IRequest<Result<RoomResponse>>;
